Listat luokka luotu, tarkoituksena helpottaa pidempien laskujen laskettavuutta.
